cleanup infra and prepare for github pages
