+++ docs/panel-example/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Hello world</title>
+    <script src="build/index.js"></script>
+    <style type="text/css">
+      body {
+        margin: 0;
+        margin-top: 50px;
+      }
+    </style>
+  </head>
+  <body>
+      <demo-app></demo-app>
+  </body>
+</html>
