+++ docs/panel-example/index.jade
