+++ docs/style-guide/colors.json
@@ -1,0 +1,1 @@
+../../stylesheets/colors.json
