+++ docs/style-guide/index.html
@@ -2,11 +2,7 @@
 <html>
     <head>
         <title>Mixpanel product style guide</title>
-        <script src="../node_modules/babel-polyfill/dist/polyfill.js"></script>
-        <script src="../node_modules/webcomponents.js/webcomponents.js"></script>
-        <script src="./build/mixpanel-common.js"></script>
-        <link rel="stylesheet" type="text/css" href="css/common.css">
-        <link rel="stylesheet" type="text/css" href="css/index.css">
+        <script src="build/index.js"></script>
     </head>
     <body class="mp-font-paragraph">
         <div class="section">
@@ -68,18 +64,18 @@
 
         <div class="section">
             <h1>Modals</h1>
-            <p><a href="#" onclick="openModal(); return false;">Open modal</a></p>
-            <p><a href="#" onclick="openPopup(); return false;">Open popup</a></p>
-            <p><a href="#" onclick="openAlertModal(); return false;">Open alert</a></p>
+            <p><a href="#" onclick="modal.open(); return false;">Open modal</a></p>
+            <p><a href="#" onclick="popup.open(); return false;">Open popup</a></p>
+            <p><a href="#" onclick="alertModal.open(); return false;">Open alert</a></p>
             <mp-modal id="modal" closeable>
                   <div class="mp-modal-title" slot-body>Easily track events</div>
                   <div class="mp-modal-subtitle" slot-body>Click on any element on your page (ex. the buy button) to create an event and access all historical data for that event.</div>
-                  <mp-button class="mp-button-modal mp-button-primary" onclick="closeModal();" slot-button arrow-next>Create your first event</mp-button>
+                  <mp-button class="mp-button-modal mp-button-primary" onclick="modal.close();" slot-button arrow-next>Create your first event</mp-button>
             </mp-modal>
             <mp-modal id="popup" width="500px" modal-type="popup" closeable>
                   <div class="mp-modal-title" slot-body>Easily track events</div>
                   <div class="mp-modal-subtitle" slot-body>Click on any element on your page (ex. the buy button) to create an event and access all historical data for that event.</div>
-                  <mp-button class="mp-button-modal mp-button-primary" onclick="closePopup();" slot-button>Create your first event</mp-button>
+                  <mp-button class="mp-button-modal mp-button-primary" onclick="popup.close();" slot-button>Create your first event</mp-button>
             </mp-modal>
             <mp-modal id="alert-modal" modal-type="popup" closeable alert>
                   <div slot-body>
@@ -87,12 +83,6 @@
                         <div class="mp-modal-alert-subtitle">If your issue persists please contact our support team at support@mixpanel.com to get help quickly</div>
                   </div>
             </mp-modal>
-            <mp-modal id="purple-modal" width="500px" closeable>
-                  <div class="mp-modal-title" slot-body>I'm a modal with a purple button</div>
-                  <div class="mp-modal-subtitle" slot-body>We use this for upsell dialogs</div>
-                  <mp-button class="mp-button-modal mp-button-purple" onclick="closePurpleModal();" slot-button arrow-next>Upgrade</mp-button>
-            </mp-modal>
         </div>
-        <script src="js/index.js"></script>
     </body>
 </html>
