+++ docs/style-guide/index.styl
@@ -1,0 +1,1 @@
+@import '../../stylesheets/common.styl'
