--- examples/colors.json
@@ -1,1 +1,0 @@
-../stylesheets/colors.json
