--- examples/panel/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Hello world</title>
-    <script src="../../node_modules/babel-polyfill/dist/polyfill.js"></script>
-    <script src="../../node_modules/webcomponents.js/webcomponents.js"></script>
-    <script src="../build/mixpanel-common.js"></script>
-    <script src="../build/panel-example.js"></script>
-    <style type="text/css">
-      body {
-        margin: 0;
-        margin-top: 50px;
-      }
-    </style>
-  </head>
-  <body>
-      <demo-app></demo-app>
-  </body>
-</html>
